MGP - var upper bound & ASE Atoms

+++ flare/kernels/mc_simple.py
@@ -835,7 +835,7 @@ def three_body_efs_self(env1: AtomicEnvironment, hyps: 'ndarray',
 
 
 def two_body_mc(env1: AtomicEnvironment, env2: AtomicEnvironment,
-                d1: float, d2: float, hyps: 'ndarray', cutoffs: 'ndarray',
+                d1: int, d2: int, hyps: 'ndarray', cutoffs: 'ndarray',
                 cutoff_func: Callable = cf.quadratic_cutoff) -> float:
     """2-body multi-element kernel between two force components.
